Move sub_device_manager_tracker.hpp out of api/tt-metallium
It's just an implementation detail of sub-device support.
