+++ tt_metal/CMakeLists.txt
@@ -84,7 +84,6 @@ target_sources(
             api/tt-metalium/semaphore.hpp
             api/tt-metalium/shape.hpp
             api/tt-metalium/sub_device.hpp
-            api/tt-metalium/sub_device_manager_tracker.hpp
             api/tt-metalium/sub_device_types.hpp
             api/tt-metalium/system_mesh.hpp
             api/tt-metalium/system_memory_manager.hpp
