+++ tt_metal/api/tt-metalium/device_impl.hpp
@@ -15,13 +15,14 @@
 #include <tt-metalium/hal_types.hpp>
 #include <tt-metalium/command_queue_interface.hpp>
 #include <tt-metalium/command_queue.hpp>
-#include <tt-metalium/sub_device_manager_tracker.hpp>
 #include <tt-metalium/sub_device_types.hpp>
+#include <tt-metalium/sub_device.hpp>
 #include <tt-metalium/trace_buffer.hpp>
 #include <tt_stl/span.hpp>
 #include <tt-metalium/program_cache.hpp>
 
 namespace tt::tt_metal {
+class SubDeviceManagerTracker;
 
 // A physical PCIexpress Tenstorrent device
 class Device : public IDevice {
