+++ tt_metal/distributed/mesh_device.cpp
@@ -11,7 +11,6 @@
 #include <mesh_device_view.hpp>
 #include <small_vector.hpp>
 #include <sub_device.hpp>
-#include <sub_device_manager_tracker.hpp>
 #include <system_mesh.hpp>
 #include <tt_metal.hpp>
 #include <algorithm>
@@ -44,6 +43,7 @@
 
 #include "tt_metal/impl/allocator/l1_banking_allocator.hpp"
 #include "tt_metal/impl/sub_device/sub_device_manager.hpp"
+#include "tt_metal/impl/sub_device/sub_device_manager_tracker.hpp"
 #include <umd/device/types/xy_pair.h>
 
 enum class CoreType;
